Fix email_backend end-to-end test setup
