+++ email_backend/data_management/generate_csv.integration-test.js
@@ -1,9 +1,11 @@
 import _ from "lodash";
 import mongoose from "mongoose";
-import { get_csv_strings } from "./generate_csv";
+
 import { connect_db } from "../src/db_utils/connect_db";
 import { make_mongoose_model_from_original_template } from "../src/db_utils/log_email_and_meta_to_db";
 import { get_templates } from "../src/template_utils";
+
+import { get_csv_strings } from "./generate_csv";
 
 //Make sure there is test data to work with
 const test_template_name = "test_template.test";
