+++ email_backend/data_management/generate_csv.js
@@ -1,10 +1,13 @@
-import { connect_db } from "../src/db_utils/connect_db";
-import { get_templates } from "../src/template_utils";
-import { make_mongoose_model_from_original_template } from "../src/db_utils/log_email_and_meta_to_db";
-import _ from "lodash";
 import fs from "fs";
-import mongoose from "mongoose";
+
 import { Parser } from "json2csv";
+import _ from "lodash";
+
+import mongoose from "mongoose";
+
+import { connect_db } from "../src/db_utils/connect_db";
+import { make_mongoose_model_from_original_template } from "../src/db_utils/log_email_and_meta_to_db";
+import { get_templates } from "../src/template_utils";
 
 const templates = get_templates();
 async function extractor() {
